Mastery Home and Manual entry is seperated using routes
