+++ src/pages/mastery/ManualEntry.tsx
@@ -1,0 +1,57 @@
+import { useState } from 'react';
+import TabViewPrimary from '@/components/reusable/tabview/TabViewPrimary';
+import Shop from './Shop';
+import MachineLine from './MachineLine';
+import Machine from './Machine';
+import Plant from './Plant';
+import Element from './Element';
+import Sensor from './Sensor';
+
+const ManualEntry = () => {
+  const [activeIndex, setActiveIndex] = useState(0);
+
+  const tabs = [
+    {
+      title: 'Plant',
+      content: <Plant />,
+      key: 'plant',
+    },
+    {
+      title: 'Shop',
+      content: <Shop />,
+      key: 'shop',
+    },
+    {
+      title: 'Machine Line',
+      content: <MachineLine />,
+      key: 'machine line',
+    },
+    {
+      title: 'Machine',
+      content: <Machine />,
+      key: 'machine',
+    },
+    {
+      title: 'Element',
+      content: <Element />,
+      key: 'element',
+    },
+    {
+      title: 'Sensor',
+      content: <Sensor />,
+      key: 'sensor',
+    },
+  ];
+
+  const handleClick = (index: number) => {
+    setActiveIndex(index);
+  };
+
+  return (
+    <>
+      <TabViewPrimary type="primary" tabs={tabs} activeIndex={activeIndex} handleClick={handleClick} />
+    </>
+  );
+};
+
+export default ManualEntry;
